requirements.txt
django
django-filter
djangorestframework
psycopg2
requests
lxml
